reorganize compute function files in `vortex-array`
Make sure each compute function implementation has its own file. Yes,
this is verbose, but also it's better to be consistent (and it also
helps with discoverability, for example "which compute functions exist
for this array type?).
