+++ vortex-array/src/arrays/extension/compute/compare.rs
@@ -4,7 +4,7 @@
 use vortex_error::VortexResult;
 
 use crate::arrays::{ConstantArray, ExtensionArray, ExtensionVTable};
-use crate::compute::{CompareKernel, CompareKernelAdapter, Operator, compare};
+use crate::compute::{self, CompareKernel, CompareKernelAdapter, Operator};
 use crate::{Array, ArrayRef, register_kernel};
 
 impl CompareKernel for ExtensionVTable {
@@ -17,7 +17,7 @@ impl CompareKernel for ExtensionVTable {
         // If the RHS is a constant, we can extract the storage scalar.
         if let Some(const_ext) = rhs.as_constant() {
             let storage_scalar = const_ext.as_extension().storage();
-            return compare(
+            return compute::compare(
                 lhs.storage(),
                 ConstantArray::new(storage_scalar, lhs.len()).as_ref(),
                 operator,
@@ -27,7 +27,7 @@ impl CompareKernel for ExtensionVTable {
 
         // If the RHS is an extension array matching ours, we can extract the storage.
         if let Some(rhs_ext) = rhs.as_opt::<ExtensionVTable>() {
-            return compare(lhs.storage(), rhs_ext.storage(), operator).map(Some);
+            return compute::compare(lhs.storage(), rhs_ext.storage(), operator).map(Some);
         }
 
         // Otherwise, we need the RHS to handle this comparison.
