+++ vortex-array/src/arrays/extension/compute/mod.rs
@@ -3,117 +3,13 @@
 
 mod cast;
 mod compare;
-
-use std::sync::Arc;
-
-use vortex_dtype::ExtDType;
-use vortex_error::VortexResult;
-use vortex_mask::Mask;
-use vortex_scalar::Scalar;
-
-use crate::arrays::ExtensionVTable;
-use crate::arrays::extension::ExtensionArray;
-use crate::compute::{
-    FilterKernel, FilterKernelAdapter, IsConstantKernel, IsConstantKernelAdapter, IsConstantOpts,
-    IsSortedKernel, IsSortedKernelAdapter, MaskKernel, MaskKernelAdapter, MinMaxKernel,
-    MinMaxKernelAdapter, MinMaxResult, SumKernel, SumKernelAdapter, TakeKernel, TakeKernelAdapter,
-    filter, is_constant_opts, is_sorted, is_strict_sorted, mask, min_max, sum, take,
-};
-use crate::{Array, ArrayRef, IntoArray, register_kernel};
-
-impl FilterKernel for ExtensionVTable {
-    fn filter(&self, array: &ExtensionArray, mask: &Mask) -> VortexResult<ArrayRef> {
-        Ok(
-            ExtensionArray::new(array.ext_dtype().clone(), filter(array.storage(), mask)?)
-                .into_array(),
-        )
-    }
-}
-
-register_kernel!(FilterKernelAdapter(ExtensionVTable).lift());
-
-impl MaskKernel for ExtensionVTable {
-    fn mask(&self, array: &ExtensionArray, mask_array: &Mask) -> VortexResult<ArrayRef> {
-        let masked_storage = mask(array.storage(), mask_array)?;
-        if masked_storage.dtype().nullability() == array.ext_dtype().storage_dtype().nullability() {
-            Ok(ExtensionArray::new(array.ext_dtype().clone(), masked_storage).into_array())
-        } else {
-            // The storage dtype changed (i.e., became nullable due to masking)
-            let ext_dtype = Arc::new(ExtDType::new(
-                array.ext_dtype().id().clone(),
-                Arc::new(masked_storage.dtype().clone()),
-                array.ext_dtype().metadata().cloned(),
-            ));
-            Ok(ExtensionArray::new(ext_dtype, masked_storage).into_array())
-        }
-    }
-}
-
-register_kernel!(MaskKernelAdapter(ExtensionVTable).lift());
-
-impl SumKernel for ExtensionVTable {
-    fn sum(&self, array: &ExtensionArray) -> VortexResult<Scalar> {
-        sum(array.storage())
-    }
-}
-
-register_kernel!(SumKernelAdapter(ExtensionVTable).lift());
-
-impl TakeKernel for ExtensionVTable {
-    fn take(&self, array: &ExtensionArray, indices: &dyn Array) -> VortexResult<ArrayRef> {
-        let taken_storage = take(array.storage(), indices)?;
-        if taken_storage.dtype().nullability() == array.ext_dtype().storage_dtype().nullability() {
-            Ok(ExtensionArray::new(array.ext_dtype().clone(), taken_storage).into_array())
-        } else {
-            // The storage dtype changed (i.e., became nullable due to nullable indices)
-            let ext_dtype = Arc::new(ExtDType::new(
-                array.ext_dtype().id().clone(),
-                Arc::new(taken_storage.dtype().clone()),
-                array.ext_dtype().metadata().cloned(),
-            ));
-            Ok(ExtensionArray::new(ext_dtype, taken_storage).into_array())
-        }
-    }
-}
-
-register_kernel!(TakeKernelAdapter(ExtensionVTable).lift());
-
-impl MinMaxKernel for ExtensionVTable {
-    fn min_max(&self, array: &ExtensionArray) -> VortexResult<Option<MinMaxResult>> {
-        Ok(
-            min_max(array.storage())?.map(|MinMaxResult { min, max }| MinMaxResult {
-                min: Scalar::extension(array.ext_dtype().clone(), min),
-                max: Scalar::extension(array.ext_dtype().clone(), max),
-            }),
-        )
-    }
-}
-
-register_kernel!(MinMaxKernelAdapter(ExtensionVTable).lift());
-
-impl IsConstantKernel for ExtensionVTable {
-    fn is_constant(
-        &self,
-        array: &ExtensionArray,
-        opts: &IsConstantOpts,
-    ) -> VortexResult<Option<bool>> {
-        is_constant_opts(array.storage(), opts)
-    }
-}
-
-register_kernel!(IsConstantKernelAdapter(ExtensionVTable).lift());
-
-impl IsSortedKernel for ExtensionVTable {
-    fn is_sorted(&self, array: &ExtensionArray) -> VortexResult<Option<bool>> {
-        is_sorted(array.storage())
-    }
-
-    fn is_strict_sorted(&self, array: &ExtensionArray) -> VortexResult<Option<bool>> {
-        is_strict_sorted(array.storage())
-    }
-}
-
-register_kernel!(IsSortedKernelAdapter(ExtensionVTable).lift());
+mod filter;
+mod is_constant;
+mod is_sorted;
+mod mask;
+mod min_max;
+mod sum;
+mod take;
 
 #[cfg(test)]
 mod test {
