+++ vortex-array/src/arrays/null/compute/mod.rs
@@ -2,70 +2,20 @@
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
 mod cast;
-
-use vortex_dtype::match_each_integer_ptype;
-use vortex_error::{VortexResult, vortex_bail};
-use vortex_mask::Mask;
-
-use crate::arrays::NullVTable;
-use crate::arrays::null::NullArray;
-use crate::compute::{
-    FilterKernel, FilterKernelAdapter, MaskKernel, MaskKernelAdapter, MinMaxKernel,
-    MinMaxKernelAdapter, MinMaxResult, TakeKernel, TakeKernelAdapter,
-};
-use crate::{Array, ArrayRef, IntoArray, ToCanonical, register_kernel};
-
-impl FilterKernel for NullVTable {
-    fn filter(&self, _array: &Self::Array, mask: &Mask) -> VortexResult<ArrayRef> {
-        Ok(NullArray::new(mask.true_count()).into_array())
-    }
-}
-
-register_kernel!(FilterKernelAdapter(NullVTable).lift());
-
-impl MaskKernel for NullVTable {
-    fn mask(&self, array: &NullArray, _mask: &Mask) -> VortexResult<ArrayRef> {
-        Ok(array.to_array())
-    }
-}
-
-register_kernel!(MaskKernelAdapter(NullVTable).lift());
-
-impl TakeKernel for NullVTable {
-    #[allow(clippy::cast_possible_truncation)]
-    fn take(&self, array: &NullArray, indices: &dyn Array) -> VortexResult<ArrayRef> {
-        let indices = indices.to_primitive();
-
-        // Enforce all indices are valid
-        match_each_integer_ptype!(indices.ptype(), |T| {
-            for index in indices.as_slice::<T>() {
-                if (*index as usize) >= array.len() {
-                    vortex_bail!(OutOfBounds: *index as usize, 0, array.len());
-                }
-            }
-        });
-
-        Ok(NullArray::new(indices.len()).into_array())
-    }
-}
-
-register_kernel!(TakeKernelAdapter(NullVTable).lift());
-
-impl MinMaxKernel for NullVTable {
-    fn min_max(&self, _array: &NullArray) -> VortexResult<Option<MinMaxResult>> {
-        Ok(None)
-    }
-}
-
-register_kernel!(MinMaxKernelAdapter(NullVTable).lift());
+mod filter;
+mod mask;
+mod min_max;
+mod take;
 
 #[cfg(test)]
 mod test {
+    use rstest::rstest;
     use vortex_buffer::buffer;
     use vortex_dtype::DType;
     use vortex_mask::Mask;
 
-    use crate::arrays::null::NullArray;
+    use crate::arrays::NullArray;
+    use crate::compute::conformance::consistency::test_array_consistency;
     use crate::compute::conformance::filter::test_filter_conformance;
     use crate::compute::conformance::mask::test_mask_conformance;
     use crate::compute::conformance::take::test_take_conformance;
@@ -119,14 +69,6 @@ mod test {
         test_take_conformance(NullArray::new(1).as_ref());
         test_take_conformance(NullArray::new(10).as_ref());
     }
-}
-
-#[cfg(test)]
-mod tests {
-    use rstest::rstest;
-
-    use crate::arrays::NullArray;
-    use crate::compute::conformance::consistency::test_array_consistency;
 
     #[rstest]
     // From test_all_consistency
