+++ vortex-array/src/arrays/varbinview/compute/mod.rs
@@ -8,6 +8,7 @@ mod is_sorted;
 mod mask;
 mod min_max;
 mod take;
+
 #[cfg(test)]
 mod tests {
     use vortex_buffer::buffer;
